Created traffic simulation test. Does not work yet.
